+++ KABUM-CLONE/EstudoSolid/src/main/java/org/example/model/dao/DatabaseConnect.java
@@ -6,8 +6,8 @@ import java.sql.DriverManager;
 public class DatabaseConnect {
 
     private static final String URL = "jdbc:mysql://localhost:3306/kabumClone";
-    private static final String LOGIN = "root";
-    private static final String PASSWORD = "jojokau8871";
+    private static final String LOGIN = "Root";
+    private static final String PASSWORD = "root";
 
 
     public static Connection obterConexao(){
